+++ CHANGES.md
@@ -12,6 +12,7 @@ For a full list of changes, see the [git commit log][log]
  * Add support to run under MSYS2 environment
  * Use CYGWRUN_POSIX_ROOT instead deprecated CYGWIN_ROOT and POSIX_ROOT
  * Set PATH to CYGWRUN_POSIX_PATH if set by user
+ * Add **-o** command option that sets PATH to ORIGINAL_PATH
 
 ## v1.2.1